Divided the module's definition of Fragment and ViewModel into each Activity's module.
