+++ app/src/main/java/io/github/droidkaigi/confsched2018/di/AppComponent.kt
@@ -23,6 +23,7 @@ import javax.inject.Singleton
     AppModule::class,
     NetworkModule::class,
     DatabaseModule::class,
+    ViewModelModule::class,
     MainActivityBuilder::class,
     MapActivityBuilder::class,
     AboutThisAppActivityBuilder::class,
