+++ app/src/main/java/io/github/droidkaigi/confsched2018/di/activitymodule/MainActivityBuilder.kt
@@ -2,15 +2,10 @@ package io.github.droidkaigi.confsched2018.di.activitymodule
 
 import dagger.Module
 import dagger.android.ContributesAndroidInjector
-import io.github.droidkaigi.confsched2018.di.FragmentBuildersModule
 import io.github.droidkaigi.confsched2018.presentation.MainActivity
 
-@Module interface MainActivityBuilder {
-    @ContributesAndroidInjector(
-            modules = [
-                FragmentBuildersModule::class,
-                MainActivityModule::class
-            ]
-    )
+@Module
+interface MainActivityBuilder {
+    @ContributesAndroidInjector(modules = [MainActivityModule::class])
     fun contributeMainActivity(): MainActivity
 }
