+++ app/src/main/java/io/github/droidkaigi/confsched2018/di/activitymodule/MapActivityBuilder.kt
@@ -2,15 +2,10 @@ package io.github.droidkaigi.confsched2018.di.activitymodule
 
 import dagger.Module
 import dagger.android.ContributesAndroidInjector
-import io.github.droidkaigi.confsched2018.di.FragmentBuildersModule
 import io.github.droidkaigi.confsched2018.presentation.map.MapActivity
 
-@Module interface MapActivityBuilder {
-    @ContributesAndroidInjector(
-            modules = [
-                FragmentBuildersModule::class,
-                MapActivityModule::class
-            ]
-    )
+@Module
+interface MapActivityBuilder {
+    @ContributesAndroidInjector(modules = [MapActivityModule::class])
     fun contributeMapActivity(): MapActivity
 }
