+++ app/src/main/java/io/github/droidkaigi/confsched2018/di/activitymodule/MapActivityModule.kt
@@ -3,8 +3,13 @@ package io.github.droidkaigi.confsched2018.di.activitymodule
 import android.support.v7.app.AppCompatActivity
 import dagger.Binds
 import dagger.Module
+import dagger.android.ContributesAndroidInjector
 import io.github.droidkaigi.confsched2018.presentation.map.MapActivity
+import io.github.droidkaigi.confsched2018.presentation.map.MapFragment
 
-@Module interface MapActivityModule {
+@Module
+interface MapActivityModule {
     @Binds fun providesAppCompatActivity(activity: MapActivity): AppCompatActivity
+
+    @ContributesAndroidInjector fun contributeMapFragment(): MapFragment
 }
