+++ app/src/main/java/io/github/droidkaigi/confsched2018/di/activitymodule/SettingsActivityBuilder.kt
@@ -2,15 +2,10 @@ package io.github.droidkaigi.confsched2018.di.activitymodule
 
 import dagger.Module
 import dagger.android.ContributesAndroidInjector
-import io.github.droidkaigi.confsched2018.di.FragmentBuildersModule
 import io.github.droidkaigi.confsched2018.presentation.settings.SettingsActivity
 
-@Module interface SettingsActivityBuilder {
-    @ContributesAndroidInjector(
-            modules = [
-                FragmentBuildersModule::class,
-                SettingsActivityModule::class
-            ]
-    )
+@Module
+interface SettingsActivityBuilder {
+    @ContributesAndroidInjector(modules = [SettingsActivityModule::class])
     fun contributeSettingsActivity(): SettingsActivity
 }
