+++ app/src/main/java/io/github/droidkaigi/confsched2018/di/activitymodule/SpeakerDetailActivityBuilder.kt
@@ -2,15 +2,10 @@ package io.github.droidkaigi.confsched2018.di.activitymodule
 
 import dagger.Module
 import dagger.android.ContributesAndroidInjector
-import io.github.droidkaigi.confsched2018.di.FragmentBuildersModule
 import io.github.droidkaigi.confsched2018.presentation.speaker.SpeakerDetailActivity
 
-@Module interface SpeakerDetailActivityBuilder {
-    @ContributesAndroidInjector(
-            modules = [
-                FragmentBuildersModule::class,
-                SpeakerDetailActivityModule::class
-            ]
-    )
+@Module
+interface SpeakerDetailActivityBuilder {
+    @ContributesAndroidInjector(modules = [SpeakerDetailActivityModule::class])
     fun contributeSpeakerDetailActivity(): SpeakerDetailActivity
 }
