+++ app/src/main/java/io/github/droidkaigi/confsched2018/di/activitymodule/TopicDetailActivityBuilder.kt
@@ -2,15 +2,10 @@ package io.github.droidkaigi.confsched2018.di.activitymodule
 
 import dagger.Module
 import dagger.android.ContributesAndroidInjector
-import io.github.droidkaigi.confsched2018.di.FragmentBuildersModule
 import io.github.droidkaigi.confsched2018.presentation.topic.TopicDetailActivity
 
-@Module interface TopicDetailActivityBuilder {
-    @ContributesAndroidInjector(
-            modules = [
-                FragmentBuildersModule::class,
-                TopicDetailActivityModule::class
-            ]
-    )
+@Module
+interface TopicDetailActivityBuilder {
+    @ContributesAndroidInjector(modules = [TopicDetailActivityModule::class])
     fun contributeTopicDetailActivity(): TopicDetailActivity
 }
